requirements.txt
asgiref==3.8.1
Django==5.1.1
sqlparse==0.5.1
psycopg2-binary==2.9.9
django-extensions==3.2.3
ipython==8.27.0
django-debug-toolbar==4.4.6
pillow==10.4.0
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
djangorestframework==3.15.2
markdown==3.7
django-filter==24.3